refactor: テストをdecodeモジュールに移動
